Rename ConfirmDeleteCatalog to ConfirmDeleteDialog
